feat: 리뷰에 s3 적용 (#44)

+++ src/main/java/ggudock/domain/review/application/ReviewService.java
@@ -2,9 +2,11 @@ package ggudock.domain.review.application;
 
 import ggudock.domain.item.entity.Item;
 import ggudock.domain.item.repository.ItemRepository;
+import ggudock.domain.picture.application.S3UploadService;
 import ggudock.domain.review.dto.ReviewRequest;
 import ggudock.domain.review.dto.ReviewResponse;
 import ggudock.domain.review.entity.Review;
+import ggudock.domain.review.entity.ReviewImage;
 import ggudock.domain.review.repository.ReviewRepository;
 import ggudock.domain.user.entity.User;
 import ggudock.domain.user.repository.UserRepository;
@@ -16,7 +18,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
@@ -29,31 +33,53 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
     private final ItemRepository itemRepository;
     private final UserRepository userRepository;
+    private final S3UploadService s3UploadService;
+    private final ReviewImageService reviewImageService;
 
     @Transactional
-    public ReviewResponse saveReview(ReviewRequest reviewRequest, String email) {
+    public ReviewResponse saveReview(ReviewRequest reviewRequest, String email, List<MultipartFile> image) throws IOException {
         User user = getUser(email);
         Item item = getItem(reviewRequest.getItemId());
         Review review = createReview(reviewRequest, item, user);
         Review savedReview = reviewRepository.save(review);
 
+        if (!image.isEmpty()) {
+            for (MultipartFile multipartFile : image)
+                reviewImageService.save(multipartFile, review);
+        }
+
         return getDetail(savedReview.getId());
     }
 
     @Transactional
-    public Review save(Review review) {
-        return reviewRepository.save(review);
-    }
+    public Long delete(Long reviewId, String email) {
+        if (checkLoginEmial(reviewId, email)) {
+            List<ReviewImage> images = reviewImageService.findImages(reviewId);
 
-    @Transactional
-    public Long delete(Long reviewId) {
-        reviewRepository.deleteById(reviewId);
+            if (!images.isEmpty()) {
+                for (ReviewImage image : images)
+                    s3UploadService.fileDelete(image.getImageUrl());
+            }
+
+            reviewImageService.deleteAll(reviewId);
+            reviewRepository.deleteById(reviewId);
+        }
         return reviewId;
     }
 
     @Transactional
     public void deleteList(String email) {
         List<Review> reviewList = reviewRepository.findByUser_EmailOrderByDateDesc(email);
+        for (Review review : reviewList) {
+            List<ReviewImage> images = reviewImageService.findImages(review.getId());
+
+            if (!images.isEmpty()) {
+                for (ReviewImage image : images)
+                    s3UploadService.fileDelete(image.getImageUrl());
+            }
+            reviewImageService.deleteAll(review.getId());
+        }
+
         reviewRepository.deleteAll(reviewList);
     }
 
@@ -151,5 +177,10 @@ public class ReviewService {
                 .build();
     }
 
+    private boolean checkLoginEmial(Long reviewId, String email) {
+        User login = userRepository.findByEmail(email);
+        Review review = reviewRepository.findById(reviewId).orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW));
+        return login.getEmail().equals(review.getUser().getEmail());
+    }
 
 }
